Document RAG System/requirements.txt
chromadb
sentence-transformers
google-generativeai
python-dotenv
tf-keras
fastapi
uvicorn[standard]
python-multipart
pypdf
python-docx